1.0.6.9-Mandatory Upgrade
- Decrease Superblock Start height to 21710

+++ src/masternode-payments.cpp
@@ -749,7 +749,7 @@ bool CMasternodeBlockPayees::IsTransactionValid(const CTransaction& txNew)
         }
     }
 
-    LogPrintf("CMasternodeBlockPayees::IsTransactionValid -- ERROR: Missing required payment, possible payees: '%s', amount: %f biblepay\n", strPayeesPossible, (float)nMasternodePayment/COIN);
+    LogPrintf("CMasternodeBlockPayees::IsTransactionValid -- ERROR: Missing required payment, possible payees: '%s', amount: %f biblepay\n", strPayeesPossible, (double)nMasternodePayment/COIN);
     return false;
 }
